requirements.txt
deepface
opencv-python
python-dotenv
requests
